Made the adam path more user-friendly by first looking into the submission folder and then in adam folder. This is important for tier > 1 datasets.
eg imagine adsl is finalized in submission/adam folder and adae is not yet created and resides in adam folder. However we need adsl for adae. By using the envsetup::read_path(adam, "adsl.xpt"), the function will first look if adsl is ready in submission/adam. If yes, then it will load this one. If no, it will take the one from the adam folder.